// src/app/shared/models/User.ts
export class User {
  id: number;
  firstName: string;
  lastName: string;
  address: string;
  cin: string;
  gender: String;
  diploma: string;
  username: string;
  email: string;
  mobile: string;
  quote: string;
  manager: boolean;
  yearsOfExperience: number;
  entryDate: any;
  integrationDate: any;
  departureDate: any;
  password: string;
  confirmPassword: string;
}